ion: add COMPAT_ION_IOC_SUNXI_FLUSH_RANGE in compat_ion_ioctl for support 32bit app.
move ION_IOC_SUNXI_FLUSH_RANGE from sunxi_ion_ioctl to ion_ioctl/compat_ion_ioctl.
also change sunxi_phys_data.phys_addr and size to unsigned int, to support both linux-3.4 and linux-3.10

+++ drivers/staging/android/uapi/ion_sunxi.h
@@ -24,8 +24,8 @@ typedef struct {
 
 typedef struct {
 	ion_user_handle_t handle;
-	unsigned long phys_addr;
-	unsigned long size;
+	unsigned int phys_addr;
+	unsigned int size;
 }sunxi_phys_data;
 
 #define ION_IOC_SUNXI_FLUSH_RANGE           5
